For now, move the method registration into a setupEngine method. This is phase 1 for allowing the library to run as separate engines
